Upgrade to latest CottonRPG

+++ src/main/java/io/github/cottonmc/skillcheck/impl/SkillCheckCharacterResource.java
@@ -7,7 +7,7 @@ import io.github.cottonmc.cottonrpg.data.resource.Ticker;
 public class SkillCheckCharacterResource extends SimpleCharacterResource {
 	private int ticksPerUnit;
 
-	public SkillCheckCharacterResource(long defaultValue, long maxValue, long unitsPerBar, int ticksPerUnit, int color, ResourceVisibility vis) {
+	public SkillCheckCharacterResource(int defaultValue, int maxValue, int unitsPerBar, int ticksPerUnit, int color, ResourceVisibility vis) {
 		super(defaultValue, maxValue, unitsPerBar, ticksPerUnit, color, vis);
 		this.ticksPerUnit = ticksPerUnit;
 	}
